Move exceptions to main package
